話す人データをFirestoreから取得するhooksを作成
