--- src/lib/useTopics.ts
@@ -1,36 +1,0 @@
-import { useCallback, useState } from "react";
-import { firebaseDB } from "../firebase";
-
-interface Topic {
-  id: string;
-  content: string;
-  timestamp: any;
-}
-
-// Topics 取得 hooks
-export const useTopics = () => {
-  const [topics, setTopics] = useState<Array<Topic>>([
-    {
-      id: "",
-      content: "",
-      timestamp: null,
-    } as Topic,
-  ]);
-
-  const getTopicsFromFirestore = useCallback(() => {
-    return firebaseDB
-      .collection("topics")
-      .orderBy("timestamp", "desc")
-      .onSnapshot((snapshot) => {
-        setTopics(
-          snapshot.docs.map<Topic>((doc) => ({
-            id: doc.id,
-            content: doc.data().topic,
-            timestamp: doc.data().timestamp,
-          }))
-        );
-      });
-  }, [firebaseDB]);
-
-  return { getTopicsFromFirestore, topics };
-};
